ajout des images et des icons

+++ src/GUI/ForumBasicController.java
@@ -6,10 +6,13 @@
 package GUI;
 
 import Entite.Question;
+import Entite.Reponse;
 import Service.QuestionService;
+import Service.ReponseService;
 
 import java.net.URL;
 import java.util.List;
+import java.util.Optional;
 import java.util.ResourceBundle;
 import javafx.beans.value.ChangeListener;
 import javafx.beans.value.ObservableValue;
@@ -22,6 +25,7 @@ import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.ListCell;
 import javafx.scene.control.ListView;
 
@@ -107,6 +111,15 @@ public class ForumBasicController implements Initializable {
                        {
                        
                    Circle c = new Circle();
+                 ImageView editIcon = new ImageView();
+                   Image editimg = new Image("/GUI/Images/edit.png") ;
+                   editIcon.setImage(editimg);
+                   
+                   ImageView deleteicon = new ImageView();
+                   Image deleteimg = new Image("/GUI/Images/rubbish.png") ;
+                   deleteicon.setImage(deleteimg);
+                   
+                              
                     
     
                     c.setCenterX(50.0);
@@ -117,18 +130,27 @@ public class ForumBasicController implements Initializable {
                          Text text = new Text (String.valueOf(item.getNbr_rep()));
                          StackPane stack = new StackPane();
                         stack.getChildren().addAll(c, text);
-                         Button edit = new Button("edit");  
-                         Button delete = new Button("delete");  
+                         Button edit = new Button();  
+                         Button delete = new Button();  
                         stack.setLayoutX(30);
                         stack.setLayoutY(30);
                         
+                        edit.setGraphic(editIcon);
+                        delete.setGraphic(deleteicon);
                         
                       VBox vbox0 = new VBox(stack , new Text("Reponse"),edit,delete);
                       vbox0.setSpacing(8);
                            delete.setOnAction((event) -> {
                           
-                          DeleteQuestion(item.getId_question());
-                          list.refresh();
+                          
+                        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+                                    alert.setTitle("Information");
+                                     alert.setContentText("Vous Voulez Vraiment supprimer cette question?  ");
+                                   Optional<ButtonType> result = alert.showAndWait();
+                                    if (result.get() == ButtonType.OK){
+                               
+                                    DeleteQuestion(item.getId_question());
+                                }  
                          
                       });
                       
@@ -199,29 +221,48 @@ public class ForumBasicController implements Initializable {
                        {
  
                        Circle c = new Circle();
+                 ImageView editIcon = new ImageView();
+                   Image editimg = new Image("/GUI/Images/edit.png") ;
+                   editIcon.setImage(editimg);
+                   
+                   ImageView deleteicon = new ImageView();
+                   Image deleteimg = new Image("/GUI/Images/rubbish.png") ;
+                   deleteicon.setImage(deleteimg);
+                   
+                              
                     
     
                     c.setCenterX(50.0);
                     c.setCenterY(125.0);
                     c.setRadius(30.0);
                     c.setFill(Paint.valueOf("#097D99"));
+                    
                          Text text = new Text (String.valueOf(item.getNbr_rep()));
                          StackPane stack = new StackPane();
                         stack.getChildren().addAll(c, text);
-
-                        Button edit = new Button("edit");  
-                         Button delete = new Button("delete");  
+                         Button edit = new Button();  
+                         Button delete = new Button();  
                         stack.setLayoutX(30);
                         stack.setLayoutY(30);
                         
+                        edit.setGraphic(editIcon);
+                        delete.setGraphic(deleteicon);
+                    
                       VBox vbox0 = new VBox(stack , new Text("Reponse"),edit,delete);
                       vbox0.setSpacing(8);
                       
-                      delete.setOnAction((event) -> {
+                         delete.setOnAction((event) -> {
                           
-                          DeleteQuestion(item.getId_question());
                           
-                          list.refresh();
+                        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+                                    alert.setTitle("Information");
+                                     alert.setContentText("Vous Voulez Vraiment supprimer cette question?  ");
+                                   Optional<ButtonType> result = alert.showAndWait();
+                                    if (result.get() == ButtonType.OK){
+                               
+                                    DeleteQuestion(item.getId_question());
+                                }  
+                         
                       });
                       
                       edit.setOnAction((event) -> {
